Detached entity cannot be deleted any more

+++ LeanMapper/Repository.php
@@ -71,11 +71,15 @@ abstract class Repository
 
 	/**
 	 * @param Entity|int $arg
+	 * @throws InvalidStateException
 	 */
 	public function delete($arg)
 	{
 		$id = $arg;
 		if ($arg instanceof Entity) {
+			if ($arg->isDetached()) {
+				throw new InvalidStateException('Cannot delete detached entity.');
+			}
 			$id = $arg->id;
 			$arg->detach();
 		}
